feat(cdk): add tests for registration
Adds tests for `start_registration` and `finish_registration`, which
test only happy paths for now. Makes `Webauthn` a generic parameter so
that we can easily control the behavior of `Webauthn` during testing.
Defines two traits to represent different phases of registration:
- `WebauthnStartRegistration`: starting passkey registration
- `WebauthnFinishRegistration`: finishing passkey registration

Enables the "test-util" feature of `aws-sdk-cognitoidentityprovider`.
Introduces the `derive_builder` crate to facilitate building a
`SharedState` for tests.