feat(cdk): respond with unauthorized error
The "/finish" action of `registration` responds with 401 unauthorized
status code when the session is missing or has expired. Also adds new
test cases to test the new behavior.

The `finish_registration` no longer returns `Error` but `ErrorResponse`
instead, when there is an error. `ErrorResponse` is a new `enum` that
can distinguish specific error responses from unhandled internal errors.
For now, it can carry `Unauthroized` error. `ErrorResponse` may be moved
to the library.

Some fields in `SharedState` now have the default values for the derived
builder.