refactor(cdk): extract test for retryable SdkError
Introduces a new trait `SdkErrorExt` which provides a function
`is_retryable` that returns if an error is retryable. It is intended to
be implemented for `SdkError<E, R>` with a specific `E`. There is the
implmenetation for `SdkError<PutItemError, R>` so far.

`finish_registration` uses it to determine if it responds with a 503
error.