feat(cdk): introduce SdkErrorExt trait in lib
`lib` introduces the `SdkErrorExt` trait that adds an extra test method
`is_retryable` which returns if an error is retryable. The trait is
intended to be implemented for `SdkError<E, R>` with a specific `E` and
arbitrary `R`. The only implementation so far is for
`SdkError<PutItemError, R>`.

`aws-smithy-runtime-api` and `aws-smithy-types` crates join the normal
dependencies.