feat(cdk): introduce ErrorResponse in lib
`lib` introduces a new enum `ErrorResponse` which helps us to
differentiate errors we what to respond with a specific status code from
those we do not care about. An `ErrorResponse` can be converted into
`Response<Body>` with the `try_into` method. Any error can be converted
into `ErrorResponse::Unhandled` with the `into` method.