refactor(cdk): make Webauthn a type parameter
`bin/user-pool-triggers` makes `Webauthn` a type parameter for the
following functions:
- `function_handler`
- `define_auth_challenge`
- `create_auth_challenge`
- `verify_auth_challenge`

The above functions except for `define_auth_challenge` require
`Webauthn` to implement either or both of the following traits:
- `WebauthnCreateAuthChallenge`: required by `function_handler` and
  `create_auth_challenge`
- `WebauthnVerifyAuthChallenge`: required by `function_handler` and
  `verify_auth_challenge`

These traits are subsets of `webauthn_rs::Webauthn`.

This change will facilitate the control of the behavior of `Webauthn`
during testing.