feat(cdk): hide Cognito trigger error details
The `function_handler` function of `bin/user-pool-triggers` returns a
generic error message instead of retruning the error object produced by
the trigger function when it fails. Because the error object might
contain sensitive information. We have to examine the CloudWatch Logs
log to see the error details.

Also adds a test case which tests the new behavior.