feat(app): show secured page if authenticated
The `SignIn` view navigates the user to the secured page (`Secured`
view) when the user is authenticated. The `SignIn` view saves the tokens
in the local storage.

The `Secured` page just shows "Secured!" if the access token is
available in the local storage, otherwise, it shows a link to the
sign-in page.