feat(cdk): make discoverable endpoint non-proxy
The integration of the discoverable start endpoint of the credentials
API becomes a non-proxy Lambda integration.

`discoverable` directly uses `lambda_runtime` instead of `lambda_http`.
It no longer needs the base path, because no path is specified when it
is invoked. It is solely the REST API's responsibility to invoke the
Lambda function from the right endpoint.