feat(cdk): make registration endpoints non-proxy
The integrations of the registration endpoints, "start" and "finish", of
the credentials API become non-proxy Lambda integrations.

`registration` directly uses `lambda_runtime` instead of `lambda_http`.
It no longer needs the base path, because no path is specified when it
is invoked. It is solely the REST API's responsibility to invoke the
Lambda function from the right endpoints. `registration` still serves
two functionalities to start or finish a registration session. It
introduces a new enum `RegistrationAction` and variants `Start` and
`Finish` which can be deserialized from objects like `{"start": ...}`
and `{"finish": ...}` respectively. Each of registration endpoints
builds a proper object by using a request mapping template.

Test cases of `registration` are also adapted to the new interface.