feat(gha): support cred registration for existing user
`finish_registration` of the `registration` app in
`lambda/authentication` checks if the Cognit user associated with the
user handle (= username) already exists in the Cognito user pool. It
creates a new Cognito user if no one is associated with the user handle,
otherwise, it uses the information of the existing user.

Also adds a new test case which tests an existing Cognit user.