fix(app): naviate to a wrong URL if base is given
Fixes the problem that viewers were navigated to a wrong URL when the
`base` Vite config was a full URL. The `base` config was directly given
to the `createWebHistory` API of Vue Router, but we should not have
specified a URL to `createWebHistory`. Extracts only the path part from
the `base` config before passing it to `createWebHistory`.