feat(cdk): add registration for a verified user
The credentials API introduces a new endpoint
`POST /registration/start-verified` that allows a verified user to
register a new credential. The endpoint invokes a new action
`StartVerified` of the `registration` app in `lambda/authentication`.
The endpoint is intended to be used by a user authenticated in a
cross-device manner to register a new device (credential).

Removes the endpoints `POST /registration/invite`, and
`POST /registration/start-invited` in favor of
`POST /registration/start-verified` which is more aligned with the
WebAuthn standard. Also removes the `Invite`, and `StartInvited` actions
from the `registration` app.

Removes comments on device invitation sessions from
`lib/session-store.ts`, because they are no longer necessary.