feat: move core resources to passquito-cdk-construct package
In a new folder `passquito-cdk-construct`:

Introduces a new TypeScript package
`@codemonger-io/passquito-cdk-construct` which defines a CDK Construct
that defines the following core Passquito resources:
- User pool (Cognito user pool)
- Credential store (DynamoDB)
- Session store (DynamoDB)
- Credentials API (API Gateway REST API)
- CloudFront distribution for apps

Separately processes and bundles the JavaScript files and the TypeScript
declaration files into the `dist` folder:
- Rollup.js for transpiling TypeScript → JavaScript, and rolling up
  JavaScript files. `rollup-plugin-esbuild` is used to transpile
  TypeScript. The default `tsconfig.json` is configured not to emit
  outputs, otherwise `rollup-plugin-esbuild` fails.
- `tsc` for generating TypeScript declaration (.d.ts) files. The
  declaration files are rolled up with `@microsoft/api-extractor`. `tsc`
  uses `tsconfig.types.json` instead of `tsconfig.json` so that
  declaration files are emitted.

Also bundles Rust source files for AWS Lambda functions
(`lambda/authentication`) into the `dist/lambda/authentication` folder.
The Rust code will be built and cached under the `node_modules` folder
of packages which install on `@codemonger-io/passquito-cdk-construct`,
but I am not sure if this works in general.

Produces the API documentation in the `api` folder with
`@microsoft/api-documenter`.

In the `cdk` folder:

Replaces the Passquito core resources with
`@codemonger-io/passquito-cdk-construct`. Since the package is installed
via a relative path, the Rust code will be built and cached directly in
the `dist` folder in the `passquito-cdk-construct` folder.