chore(passquito-cdk-construct): introduce build:noclean script
Extracts the core build process (`build:js` and `build:dts`) from the
`build` script as a `build:noclean` script. The `build` script calls the
`build:noclean` script. The script is introduced, because removing the
`dist` folder is very expensive during development; it ends up with
fresh rebuild of the Rust code.