feat(gha): publish developer package of passquito-client-js
Introduces a new GitHub Actions workflow `publish-passquito-client-js`
which publishes a developer package to the npm registry managed by
GitHub Packages. The `publish-developer-packages` calls it when commits
are pushed to the `main` branch.

As `passquito-client-js` depends on a developer release of Buefy, you
have to configure the `PACKAGE_READER_PAT` secret with a PAT which
has at least the `read:packages` scope.