feat(gha): run passquito-cdk-construct tests
Introduces a new GitHub Actions workflow `test-passquito-cdk-construct`
which runs tests of `passquito-cdk-construct` whenever commits are
pushed to any branch except for `main` or a PR is made to the `main`
branch. The `main` branch is excluded from the push trigger, because the
workflow that publishes a developer package should run the tests prior
to publishing the package.